+++ packages/cli/src/commands/template/buildWithProxy.ts
@@ -18,15 +18,15 @@ export async function buildWithProxy(
 ) {
   if (!userConfig?.dockerProxySet) {
     console.log(
-      'There was an issue during Docker authentication. Please follow the workaround steps from https://app.grotte.parallactic.fr/docs/troubleshooting/templates/build-authentication-error and then continue.'
+      'There was an issue during Docker authentication. Please follow the workaround steps from https://app.grotte.dev/docs/troubleshooting/templates/build-authentication-error and then continue.'
     )
     const yes = await confirm(
-      'Have you completed the steps from the https://app.grotte.parallactic.fr/docs/troubleshooting/templates/build-authentication-error workaround guide?'
+      'Have you completed the steps from the https://app.grotte.dev/docs/troubleshooting/templates/build-authentication-error workaround guide?'
     )
 
     if (!yes) {
       console.log(
-        'Please follow the workaround steps from https://app.grotte.parallactic.fr/docs/troubleshooting/templates/build-authentication-error and then try again.'
+        'Please follow the workaround steps from https://app.grotte.dev/docs/troubleshooting/templates/build-authentication-error and then try again.'
       )
       process.exit(1)
     }