+++ packages/python-sdk/grotte/sandbox/sandbox_api.py
@@ -76,7 +76,7 @@ class SandboxNetworkOpts(TypedDict):
     mask_request_host: NotRequired[str]
     """
     Allows specifying a custom host mask for all sandbox requests.
-    Supports ${PORT} variable. Defaults to "${PORT}-sandboxid.grotte.parallactic.fr".
+    Supports ${PORT} variable. Defaults to "${PORT}-sandboxid.grotte.dev".
 
     Examples:
     - Custom subdomain: `"${PORT}-myapp.example.com"`